requirements.txt
selenium
webdriver_manager
pandas
supabase
python-dotenv